Moved Renderer.symbols into own registry.
