+++ ts/Core/Renderer/SVG/SVGRenderer.ts
@@ -31,10 +31,7 @@ import type PositionObject from '../PositionObject';
 import type SVGAttributes from './SVGAttributes';
 import type SVGPath from './SVGPath';
 import type SVGRendererLike from './SVGRendererLike';
-import type {
-    SymbolFunction,
-    SymbolOptions
-} from './SymbolFunction';
+import type SymbolOptions from './SymbolOptions';
 
 import AST from '../HTML/AST.js';
 import Color from '../../Color/Color.js';
@@ -55,6 +52,7 @@ import Palette from '../../Color/Palette.js';
 import RendererRegistry from '../RendererRegistry.js';
 import SVGElement from './SVGElement.js';
 import SVGLabel from './SVGLabel.js';
+import Symbols from './Symbols.js';
 import TextBuilder from './TextBuilder.js';
 import U from '../../Utilities.js';
 const {
@@ -82,41 +80,6 @@ const {
  * */
 
 let hasInternalReferenceBug: (boolean|undefined);
-
-/* *
- *
- *  Functions
- *
- * */
-
-const roundedRect: SymbolFunction = function (x, y, w, h, options): SVGPath {
-    const r = (options && options.r) || 0;
-    return [
-        ['M', x + r, y],
-        ['L', x + w - r, y], // top side
-        ['C', x + w, y, x + w, y, x + w, y + r], // top-right corner
-        ['L', x + w, y + h - r], // right side
-        ['C', x + w, y + h, x + w, y + h, x + w - r, y + h], // bottom-rgt
-        ['L', x + r, y + h], // bottom side
-        ['C', x, y + h, x, y + h, x, y + h - r], // bottom-left corner
-        ['L', x, y + r], // left side
-        ['C', x, y, x, y, x + r, y] // top-left corner
-    ];
-};
-
-// #15291
-const rect: SymbolFunction = function (x, y, w, h, options): SVGPath {
-    if (options && options.r) {
-        return roundedRect(x, y, w, h, options);
-    }
-    return [
-        ['M', x, y],
-        ['L', x + w, y],
-        ['L', x + w, y + h],
-        ['L', x, y + h],
-        ['Z']
-    ];
-};
 
 /* *
  *
@@ -2124,7 +2087,7 @@ interface SVGRenderer extends SVGRendererLike {
     Element: typeof SVGElement;
     SVG_NS: string;
     escapes: Record<string, string>;
-    symbols: Record<string, SymbolFunction>;
+    symbols: typeof Symbols;
     draw: Function;
 }
 extend(SVGRenderer.prototype, {
@@ -2167,234 +2130,7 @@ extend(SVGRenderer.prototype, {
      * @name Highcharts.SVGRenderer#symbols
      * @type {Highcharts.SymbolDictionary}
      */
-    symbols: {
-        circle: function (x, y, w, h): SVGPath {
-            // Return a full arc
-            return this.arc(x + w / 2, y + h / 2, w / 2, h / 2, {
-                start: Math.PI * 0.5,
-                end: Math.PI * 2.5,
-                open: false
-            });
-        },
-
-        rect,
-
-        square: rect, // #15291
-
-        triangle: function (x, y, w, h): SVGPath {
-            return [
-                ['M', x + w / 2, y],
-                ['L', x + w, y + h],
-                ['L', x, y + h],
-                ['Z']
-            ];
-        },
-
-        'triangle-down': function (x, y, w, h): SVGPath {
-            return [
-                ['M', x, y],
-                ['L', x + w, y],
-                ['L', x + w / 2, y + h],
-                ['Z']
-            ];
-        },
-        diamond: function (x, y, w, h): SVGPath {
-            return [
-                ['M', x + w / 2, y],
-                ['L', x + w, y + h / 2],
-                ['L', x + w / 2, y + h],
-                ['L', x, y + h / 2],
-                ['Z']
-            ];
-        },
-        arc: function (x, y, w, h, options): SVGPath {
-            const arc: SVGPath = [];
-
-            if (options) {
-                const start = options.start || 0,
-                    rx = pick(options.r, w),
-                    ry = pick(options.r, h || w),
-                    proximity = 0.001,
-                    fullCircle = (
-                        Math.abs((options.end || 0) - start - 2 * Math.PI) <
-                        proximity
-                    ),
-                    // Substract a small number to prevent cos and sin of start
-                    // and end from becoming equal on 360 arcs (related: #1561)
-                    end = (options.end || 0) - proximity,
-                    innerRadius = options.innerR,
-                    open = pick(options.open, fullCircle),
-                    cosStart = Math.cos(start),
-                    sinStart = Math.sin(start),
-                    cosEnd = Math.cos(end),
-                    sinEnd = Math.sin(end),
-                    // Proximity takes care of rounding errors around PI (#6971)
-                    longArc = pick(
-                        options.longArc,
-                        end - start - Math.PI < proximity ? 0 : 1
-                    );
-
-                arc.push(
-                    [
-                        'M',
-                        x + rx * cosStart,
-                        y + ry * sinStart
-                    ],
-                    [
-                        'A', // arcTo
-                        rx, // x radius
-                        ry, // y radius
-                        0, // slanting
-                        longArc, // long or short arc
-                        pick(options.clockwise, 1), // clockwise
-                        x + rx * cosEnd,
-                        y + ry * sinEnd
-                    ]
-                );
-
-                if (defined(innerRadius)) {
-                    arc.push(
-                        open ?
-                            [
-                                'M',
-                                x + innerRadius * cosEnd,
-                                y + innerRadius * sinEnd
-                            ] : [
-                                'L',
-                                x + innerRadius * cosEnd,
-                                y + innerRadius * sinEnd
-                            ],
-                        [
-                            'A', // arcTo
-                            innerRadius, // x radius
-                            innerRadius, // y radius
-                            0, // slanting
-                            longArc, // long or short arc
-                            // Clockwise - opposite to the outer arc clockwise
-                            defined(options.clockwise) ? 1 - options.clockwise : 0,
-                            x + innerRadius * cosStart,
-                            y + innerRadius * sinStart
-                        ]
-                    );
-                }
-                if (!open) {
-                    arc.push(['Z']);
-                }
-            }
-
-            return arc;
-        },
-
-        /**
-         * Callout shape used for default tooltips, also used for rounded
-         * rectangles in VML
-         */
-        callout: function (x, y, w, h, options): SVGPath {
-            const arrowLength = 6,
-                halfDistance = 6,
-                r = Math.min((options && options.r) || 0, w, h),
-                safeDistance = r + halfDistance,
-                anchorX = options && options.anchorX,
-                anchorY = options && options.anchorY || 0;
-
-            const path = roundedRect(x, y, w, h, { r });
-
-            if (!isNumber(anchorX)) {
-                return path;
-            }
-
-            // Anchor on right side
-            if (x + anchorX >= w) {
-
-                // Chevron
-                if (
-                    anchorY > y + safeDistance &&
-                    anchorY < y + h - safeDistance
-                ) {
-                    path.splice(
-                        3,
-                        1,
-                        ['L', x + w, anchorY - halfDistance],
-                        ['L', x + w + arrowLength, anchorY],
-                        ['L', x + w, anchorY + halfDistance],
-                        ['L', x + w, y + h - r]
-                    );
-
-                // Simple connector
-                } else {
-                    path.splice(
-                        3,
-                        1,
-                        ['L', x + w, h / 2],
-                        ['L', anchorX, anchorY],
-                        ['L', x + w, h / 2],
-                        ['L', x + w, y + h - r]
-                    );
-                }
-
-            // Anchor on left side
-            } else if (x + anchorX <= 0) {
-
-                // Chevron
-                if (
-                    anchorY > y + safeDistance &&
-                    anchorY < y + h - safeDistance
-                ) {
-                    path.splice(
-                        7,
-                        1,
-                        ['L', x, anchorY + halfDistance],
-                        ['L', x - arrowLength, anchorY],
-                        ['L', x, anchorY - halfDistance],
-                        ['L', x, y + r]
-                    );
-
-                // Simple connector
-                } else {
-                    path.splice(
-                        7,
-                        1,
-                        ['L', x, h / 2],
-                        ['L', anchorX, anchorY],
-                        ['L', x, h / 2],
-                        ['L', x, y + r]
-                    );
-                }
-
-            } else if ( // replace bottom
-                anchorY &&
-                anchorY > h &&
-                anchorX > x + safeDistance &&
-                anchorX < x + w - safeDistance
-            ) {
-                path.splice(
-                    5,
-                    1,
-                    ['L', anchorX + halfDistance, y + h],
-                    ['L', anchorX, y + h + arrowLength],
-                    ['L', anchorX - halfDistance, y + h],
-                    ['L', x + r, y + h]
-                );
-
-            } else if ( // replace top
-                anchorY &&
-                anchorY < 0 &&
-                anchorX > x + safeDistance &&
-                anchorX < x + w - safeDistance
-            ) {
-                path.splice(
-                    1,
-                    1,
-                    ['L', anchorX - halfDistance, y],
-                    ['L', anchorX, y - arrowLength],
-                    ['L', anchorX + halfDistance, y],
-                    ['L', w - r, y]
-                );
-            }
-
-            return path;
-        }
-    },
+    symbols: Symbols,
 
     /**
      * Dummy function for plugins, called every time the renderer is updated.
